requirements.txt
alembic==0.8.9
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
blinker==1.4
cffi==1.9.1
click==6.7
cssmin==0.2.0
decorator==4.0.10
enum34==1.1.6
Flask==1.0.2
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.1
Flask-JSGlue==0.3.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Moment==0.6.0
flask-msearch==0.1.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.1
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
pathlib2==2.1.0
peewee==2.9.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.2.0
psycopg2==2.7.5
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.1.3
pyparsing==2.2.0
python-editor==1.0.3
python-slugify==1.2.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
Unidecode==0.4.20
webassets==0.12.1
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.1